migrate tests & delete ReportPage
